+++ NEWS.md
@@ -1,3 +1,9 @@
+### cholera 0.8.0.9368
+
+- change "case.location" argument to 'location' with values "nominal" or 
+  "orthogonal" in addNeighborhoodCases().
+  
+  
 ### cholera 0.8.0.9367
 
 - change "case.location" argument to 'location' with values "nominal" or